Added author semantic link to page template.
Depending on which blog post you read, this might enable Google to pick up
author bio and picture and integrate in search results.

For reference, here's the official documentation:
http://support.google.com/webmasters/bin/answer.py?hl=en&answer=1408986&topic=2371375&ctx=topic

However, I'm here using a <link> tag instead of an <a> tag, as it seems
that it works for other people.