Added a non-breaking space to analytics,
because lots of troubleshooting seems to indicate that if I add an sort of
character between {% when "google" %} and {% include
JB/analytics-providers/google %}, Jekyll is actually going to include the
include file, but if there's pure white space, it's not(!).

At least, this is the behaviour on my local machine, when I run

jekyll serve --safe

Now, in order to troubleshoot further, I'm committing this to master and
pushing to origin/master to see if that is also the case on blog.ploeh.dk.