Corrected comment dates,
because, as I'm writing this, 2014-08-26 22:30 UTC hasn't happened yet.
I'm basing the reader comment date and time on the premise that the time
is overall correct, but the date is off-by-one. This seems to fit the time
the pull request arrived.

Also, it looks as though I can't get my own comment dates right, either :$