Correct timestamp of reader comment.
The time used in the comment would put the time of the comment AFTER the
pull request arrived. Adjusting it back one hour fits the time of the pull
request within a couple of minutes.

People often forget that UTC is never on Daylight Saving Time, so this
time of year, Western Europe is 2 hours ahead of UTC.