Ignored .jekyll-metadata,
as recommended by the official documentation:

"This helps Jekyll keep track of which files have not been modified since
the site was last built, and which files will need to be regenerated on
the next build. This file will not be included in the generated site. Its
probably a good idea to add this to your .gitignore file."

- http://jekyllrb.com/docs/structure/