Experimenting with setting ASSET_PATH to false,
like it was before. According to the discussion at
https://github.com/plusjade/jekyll-bootstrap/issues/295#issuecomment-182503739
the changes I've previously made to archive_path, etc. might be enough. I
can't say that I understand this, but we'll see.

It looks OK on my machine, but it also did that before, and still looked
horrible in GitHub pages.