Benefit of `tryAcceptComposition`?
Hi, Thank you for those blog posts. I also read your other posts on ports and adapters and the proposed architecture makes sense in terms of how it works, but I struggle to see the benefit in a real world application. Let me maybe explain with a quick example.

In the 2nd blog post you demoed this function:

```
// int -> (DateTimeOffset -> Reservation list) -> (Reservation -> int) -> Reservation
// -> int option
let tryAccept capacity readReservations createReservation reservation =
    let reservedSeats =
        readReservations reservation.Date |> List.sumBy (fun x -> x.Quantity)
    if reservedSeats + reservation.Quantity <= capacity
    then createReservation { reservation with IsAccepted = true } |> Some
    else None
```

If I understand it correctly this function is pure if `readReservations` and `createReservation` are both pure as well and it is impure otherwise.

The benefit of having a pure function is because it is extremely easy to understand, test and reason about, that makes sense.

So then in the 3rd blog post you showed how to make `tryAccept` pure, by removing the function dependencies and replacing it with values:

```
// int -> Reservation list -> Reservation -> Reservation option
let tryAccept capacity reservations reservation =
    let reservedSeats = reservations |> List.sumBy (fun x -> x.Quantity)
    if reservedSeats + reservation.Quantity <= capacity
    then { reservation with IsAccepted = true } |> Some
    else None
```

However, additionally you had to move the impure code into a new function:

```
// Reservation -> int option
let tryAcceptComposition reservation =
    reservation.Date
    |> DB.readReservations connectionString
    |> flip (tryAccept 10) reservation
    |> Option.map (DB.createReservation connectionString)
```

So after all the application hasn't reduced the total number of impure functions (still 3 in each case - readReservations, createReservation and tryAccept[Composition]). The only difference is that one impure function `tryAccept` has been refactored into one pure function and a second impure function. What is the benefit of that? If I still remain with the same amount of impure functions and all of them are 100% testable from a unit testing point of view, did I win anything by making that extra split? I find the original `tryAccept` much easier to read and understand than the combination of `tryAccept` + `tryAcceptComposition`.

Am I missing something?