Clarified all and any monoids
The original sentence caused confusion because it looked like I was
claiming that `and` is the identity for 'all', and `or` is the identity
for 'all'. This is wrong, and it was never my intent, but I can see how
the sentence can be read that way.

Hopefully, this edit makes it clearer what I meant.