Updated title of next article
I may decide to change it again, as I'm not entirely happy with it as it
is now. However, I've found out that I can cover some other patterns
than just the structural patterns, while, on the other hand, I don't
expect to be able to cover all the structural patterns. Thus, the title
shouldn't include 'structural', as that's too narrow.