Updated overview article about specific patterns
I've struggled with giving this article a good name, but have settled on
the name it now has.

The list of specific patterns covered may look unimpressive, but my
intent is to expand it to more patterns, once I have the supporting
material in place. This article will never cover all the GoF patterns,
but I hope to be able to cover all the useful ones.