Corrected 'are monoids' to 'form monoids'
It turns out that it has been surprisingly difficult for me to keep a
consistent and correct language relating to what monoids are, and how
they relate to types. The confusion partially has its root in Haskell's
implementation of monoid as a type class, which implies that in some
sense, types and monoids are related.

This isn't entirely wrong from the perspective of Category Theory, where
one interpretation of what an 'object' means in a programming context is
exactly that objects are types.

Still, a monoid is a triple of the following: A binary operation.
Associativity. Identity.

I try to be careful when choosing my words, particularly in a C# context
without type classes, but I sometimes slip.