Polished markup of reader comment
It seems that most comment pull requests I get are weirdly formatted,
which, is my hypothesis, could be because the example I point to in my
'Wish to comment?' documentation is formatted like that. My mistake, in
other words.

I changed the comment time to reflect the time the pull request was
received.