Converted reader comments <br> tags to <p>
I usually don't touch the contents of reader comments, since the
comments are the readers', and not mine. I usually extend this principle
to formatting as well.

All comments made before the change to Jekyll, however, were made using
the form that was included with dasBlog. Thus, readers who wrote
comments using that form didn't have full control of the HTML they
submitted. I therefore consider it reasonable to edit the HTML
formatting of those comments. Particularly, something as innocuous as
changing <br> to <p> tags is something that doesn't change the
sentiment or message of individual comments.

The motivation for this particular commit is that new users of the blog
use this particular page as a template for how to comment (because
that's what I've told them to do in the README.md). I therefore want to
present a template that looks as much like what I'd like to see as
possible.