Converted reader comment time to UTC
This blog keeps time exclusively in UTC, mostly to simplify my future
work when I have to migrate the contents to a new platform.

Not that I have any specific plans to do that, but since I've already
been through that exercise once, and since I've been blogging for
thirteen years with no plans to stop, I expect it to have to happen
sooner or later. I take the long view on this...