Changed site production URL to HTTPS
This is an experiment. It's a little hard to test on my local machine
how this'll look once in production, so I'll try to make the change and
see what happens. Ready to roll back if everything turns out to be
messed up.

This is intended to be the first step in enforcing HTTPS for the site,
as tracked by the GitHub issue #430
(https://github.com/ploeh/ploeh.github.com/issues/430).