Add Elm as an example functional language
Hi Mark

Opening Prs is the way to make comments, but I thought I might as well just propose the actual change instead.

I believe that one of your aims is to promote staticcally strongly typed functional languages, and I believe that Elm is probably the best way to do this.

Elm is targeted at front end development, and makes it easy to create websites, which is what most beginners are trying to do. The language has a strong focus on simplicity and a single recommended way of doing things, which is a refreshing change, especially in the front end world. It sits very well in it's target niche, and the runtime essentially provides an impure abstract syntax tree interpreter, leaving all your code 100% pure / functional. There is also a strong and vibrant community, which is both innovative and supportive to newcomers. It doesn't have all the features of haskell, which obviously has some good points and some bad points

Thanks

Cedd