Update 2018-04-09-coalescing-composite-as-a-monoid.html
Hi Mark,

First I want to explain why I'm writing the comment here: while your blog is fascinating, I find the need for using GitHub pull requests for sending comments very combersome, especially via mobile. I hope you'd understand and edit the comment yourself or direct me to an easier way to post my comment...

And now for the comment itself:
The use of `First<T>` as a mean to make the `Read` method return a monoid is very elegant, but I'm missing something: I don't see a way to extract the actual value out of it... And if there's no value, then you end up in the same situation as before.

Maybe `First<T>` should have another method that takes an `Action<T>` and calls it only if it has a value, but who should call this method and when? 

Can you clarify how one should get the value out of `First<T>`?

Thanks in advance,
Arnon.