Remove duplicate line
I'm not a hundred percent sure, but I think that the duplicated line is
an artefact of the way Git Bash interacts with the command window. When
you issue a `git show` command, the output is piped to a Less session.
I've noticed that when I navigate the Less session (with j, k, space,
etc. keys) lines are sometimes duplicated on the screen. Or at least,
this may happen when I subsequently scroll the window after exiting the
Less session, scrolling using the Command window's own buffer.

Whether this is an issue specific to my environment or a more general
issue, I'm not aware. It's not much of an issue normally, so I haven't
investigated.

I admit that for the blog post, I just marked the blocks in my Command
window and copied them. Then I manually added the colors. A more correct
approach would probably have been to pipe the output of `git show` to a
file.

This resolves #667.