Remove author link from comment
The purpose of the author link is to enable other readers that come by
the blog page to identify the author of the comment, if the commenter so
chooses.

Linking to a GitHub repository doesn't fulfil that purpose. Other
readers who come across a comment in the future will not be able to
learn more about the author by following that link.

If a commenter wants to stay anonymous, or if a commenter has no
personal web site that he or she wishes to link to, then that's fine
with me. As the phrase in the newspaper business goes, 'the writer
wishes to remain anonymous, but his or her identity is known by the
editors'.