Correct comment metadata
Removal of dashes in the GUID is only for my own sanity's sake. I
foresee that one day I may have to port the blog away from Jekyll to
another system, and perhaps this new system will have a more appropriate
commenting system. This is likely to entail a lot of scripting on my
part, migrating articles and comments over. The more consistent I can
keep the metadata, the easier it may be to migrate...