Delete (semantically) a sentence
As described in the previous commit's response to a reader comment, the
deleted sentence is most likely incorrect.

As usual, I use semantic deletion instead of actual deletion of content,
so that readers are in a better position to understand how the text has
changed over the years.

The timestamp on the change is ridiculously over-precise. I don't think
it's necessary to include fractions of a second, but I was lazy and just
used a template I had lying around. The time data comes from a call to
`DateTime.Now` in .NET. Obviously, seconds and even minutes have now
passed since then, just by typing in this commit message...