Remove licence file
The source code was licenced with the MIT licence. This is hardly
consistent with the copyright notice that I use on the blog itself.

The MIT licence gives anyone the right to redistribute and reuse the
entire repository of the blog, even for commercial use. I don't think
that it's fair that other people should (even hypothetically) be able to
make money on my work. I do, morally, reserve that right for myself,
although I have no plans to do so. The blog, so far, has been my gift to
the world.

Readers are still welcome to fork the repository in order to send
comments in the form of pull requests.

I do acknowledge that part of the repository is the work of others,
particularly the Jekyll and Bootstrap code. Removing the licence file
doesn't change that.

As I've tried to outline, I don't plan to commercialise the blog, and
even if I do, affected would be the contents, not the technology used to
host it. Removing the licence is only done to make it a little more
difficult for others to redistribute the contents and pass it off as
their own work.

Readers, commentors, professionals, scholars, etc. are still welcome to
read and use the blog. Quoting from it is still allowed with proper
citation, and within normal 'fair use' limits.