Replace <br/> with <br>
This is an effort to have more consistent markup, also related to
ongoing work with adding local anchors to comments:
https://github.com/Danielovich/PloehComments/issues/5

If anyone's wondering why I don't just make a search-and-replace, I want
to visually verify that these changes don't mess anything up - at least
for a few posts.

If you're a front-end developer and find this ridiculous, then please
have patience with me. I rarely deal with such issues these days, and I
haven't kept up with web page rendering or HTML standards.

This one is a bit different than the previous ones, since I've replaced
most of the `<br/>` with actual line breaks, since they appeared in
`<pre>` tags. There was a brief period where it turned out that I'd
misconfigured the Copy-Source-As-HTML Visual Studio add-in so that it'd
add `<br/>` tags instead of real line breaks.