Remove illegal HTML tag from reader code example
While I completely understand what was meant, and the more proper way to
address this issue might be to HTML-escape the opening and closing angle
brackets, I consider reader content off-limits to my interpretations.
In this case the interpretation might be straightforward, but there
might be other cases were it isn't so.

My position is to preserve render-fidelity, so to speak. What other
readers have (potentially) been seeing for years should be preserved.