Add clickable anchors to all comments
This commit represents the verbatim changes made by the code in
https://github.com/Danielovich/PloehComments. It contains a number of
errors that are not the fault of the tool, but a result of markup errors
in the documents. Therefore, this commit SHOULD NOT BE MERGED until
future commits have fixed the issues.

I am, however, making this as an individual commit so that I have a way
to trace the origin of a problem, in case I've missed one originally.