Correct file rename error
I'd started writing the article yesterday, which explains why the file
was originally named with the `2023-11-12` prefix. Althgough I renamed
the file from the command line, I forgot that I still had it open in my
editor. When I changed the metadata in the open file and saved it, the
editor created the 'old' file again.