Precise:  Serialization w/ & without Reflection
Haskell has reflection through [Typeable](https://hackage.haskell.org/package/base-4.19.0.0/docs/Data-Typeable.html#t:Typeable), however it's not widely use.

I tend to use it at the boundaries of my Domain, when I have runtime types uncertainty.

I agree that they are not used for serialization (as far as I know), and that the RTS does not have enough knowledge on the types it deals with (and we cannot access to what it knows).