Fixing the typo in traversals post https://blog.ploeh.dk/2024/11/11/traversals/
Fixing the typo in traversals post

Post URL:
https://blog.ploeh.dk/2024/11/11/traversals/

Actual text:
Since tasks (e.g. Task<T>) are applicative functors (the are, because they are monads, and all monads are applicative functors), that second requirement is fulfilled for the above example.

Fixed text:
Since tasks (e.g. Task<T>) are applicative functors (they are, because they are monads, and all monads are applicative functors), that second requirement is fulfilled for the above example.