Add cspell.json file
I've always known that when I write source code, I use American
spelling, as this is, for most programming languages, the norm. The more
I use VS Code for programming, the more the default setting of the spell
checker to UK English is a bother.

Finally realising that the blog is the exception, I've changed the
default spelling language in VS Code to American English, which then
entails that I need a custom settings file for this particular
repository.